react: Add test for dependencies changing but value not changing
This is important because the README specifies that callbacks are called
"when the cell's value in a new stable state **has changed** from the
previous stable state."

The example solution already passes this test, so no change to the
example solution is necessary.

However, there is a type of solution that will not pass this test, so
this test is indeed still useful:
This is the type of solution that, when a cell gets updated, marks it
dirty, and then afterwards calls the callbacks of all dirty cells.
This test shows that a dirty bit is not enough information; you need to
store the last stable value of the cell so that you can compare it
against the new stable value.