Travis: Add configlet fmt
This commit helps to catch bad formatting in the track's JSON files.
Such problems now cause the build to fail.

The `configlet fmt` command checks that:
- `config.json` and `maintainers.json` have consistent indentation.

- Exercise topics in `config.json` are alphabetized and normalized.

If not, the command fixes the formatting and exits with exit code 0.
This causes `git diff --exit-code` to show the diff and exit with exit
code 1.

Decisions:
- Place `fmt` after `lint`, which does more important things. Note that
  they both check that `config.json` and `maintainers.json` have
  consistent JSON syntax, and if not, output the first problem found and
  exit with exit code 1.