Refactor(twelve-days): Workaround compiler bug
This commit is to avoid a Nim compiler bug that is seen when using the
new `check_exercises.nim` (see the next commit).

It is a useful minor feature to set the default value of one parameter
to that of another. However, it currently only works when the function
is called from the top level. This interferes with the new
`check_exercises.nim`, which wraps all the tests for an exercise inside
a proc to avoid a "too many global variables" error when joining many
top-level `unittest` tests.