READMEs: Improve style of hints
Hint files should start with a header so that it is easy to see where
the track-specific content begins.

This commit adds a header where it was missing and gives hints a
consistent style.

It also adds a newline to the `grade-school` README to make it
consistent with the output of `configlet generate`. This fixes an
oversight from #158.