Refactor(robot-name): Fix unused import warning
This commit stops a warning from being printed when compiling the tests
for `robot-name` and clarifies how to enable its bonus test.

Background:
There is a bonus test for `robot-name` that was commented out because:
- The test is better left opt-in.
- The test requires more computation than others on the Nim track.
- The `skip` functionality in `unittest` still executes the "skipped"
  test code.

But the otherwise unneeded `sets` import was left uncommented because:
- It made it easier for users to enable the bonus test.
- Modules cannot be imported where the bonus test is - `import` is only
  allowed at top level.

This seemed reasonable when the exercise was implemented, but later
versions of the Nim compiler print this warning for the unused import:

```
test_robot_name.nim(1, 18) Warning: imported and not used: 'sets'
```

The commits that added unused import detection to the Nim compiler are:
- https://github.com/nim-lang/Nim/commit/a5d6080
- https://github.com/nim-lang/Nim/commit/114da04

Some options to address the warning:
1) Remove the bonus test and the import. This is unsatisfying because
   the constraint from the bonus test is arguably the most interesting
   and unique part of the exercise.
2) Uncomment the bonus test. This is bad because it wouldn't be opt-in.
3) Keep the bonus test commented out and also comment out the import.
   This is better, but the user who wants to run the bonus test would
   need to uncomment in two places.
4) Use `when` to conditionally import `sets` and run the test.
5) Use `when` to conditionally run the test, and refactor the bonus test
   to not require an import.

This commit implements option 4 and prepares for option 5. The warning
is no longer printed and the test is still easy to enable.