Refactor(robot-name): Optimize bonus test
This commit improves the performance of the `robot-name` bonus test. See
the next commit for benchmarks.

This commit also allows users to implement the name as `array[5, char]`
instead of `string` if they prefer. It seems reasonable to support
possible optimizations here because `robot-name` (with the bonus test)
is by far the most computationally taxing exercise on the Nim track.

Also note:
- `robot-name` is currently the only exercise on the Nim track that
  lacks upstream `canonical-data.json`.
- Most tracks that implement `robot-name` do not test generating all the
  names.

Decisions:
- Use a `seq` rather than a `HashSet` (with or without custom hashing)
  or an `Intset`.
- Count the duplicate names and `check` that, rather than having a
  `check` for each name (which prints too many lines to the terminal
  for a failing implementation). This approach is also more
  user-friendly than using the `unittest` variable `abortOnError` or the
  `fail` template, which do not print the usual red `[FAILED]`.
- Similarly, count the invalid names. This has little overhead and also
  ensures that `toInt` always generates a valid index for the `seq`.