Refactor(robot-name): Optimize example for bonus test
This commit better optimizes the `robot-name` example implementation for
the case of generating a large number of names. Together, the changes in
this commit and the previous one make the now-enabled bonus test finish
about 7.5x faster (for a debug build).

This also speeds up running the tests for all exercises by a similar
factor as `robot-name` dominated the `./all_tests` running time.

The table below shows benchmarks for the `robot-name` test suite. Each
number is the mean performance (200 runs) relative to a debug build from
2 commits ago, which took about 2 seconds on a typical laptop. "10x"
means "finished in a tenth of the time".

| Commit                     | debug | -d:release | -d:danger | + flto |
|----------------------------+-------+------------+-----------+--------|
| 2 commits ago              |  1.0x |       5.3x |      5.9x |  10.1x |
| 1 commit ago               |  1.2x |       7.4x |      8.7x |  23.9x |
| This commit, no {.inline.} |  7.5x |      56.8x |     61.9x | 103.8x |
| This commit                |  7.5x |      72.1x |     95.5x | 102.1x |
| This commit, no bonus test | 24.3x |     125.2x |    125.8x | 129.2x |

System details:
- x86_64
- Linux
- Nim devel 1.1.1 (git hash 347a046, 2019-12-13)
- GCC 9.2.0

The final column corresponds to `nim c -d:danger --passC:-flto`.

The final row shows that for optimized builds the running time is now
mostly spent shuffling names, rather than running the bonus test itself.
Shuffling takes about 70% of the running time when the bonus test is
enabled, or 90% when the bonus test is disabled.

Note that, even after this commit, testing `robot-name` still takes
twice as long as all the other exercises combined (for a debug build).