CI: Cache the Nim installation
The runtime of `bin/install-nim.sh` on Travis CI was typically about:
- 5 minutes for amd64 + Linux
- 4 minutes for amd64 + macOS
- 7 minutes for arm64 + Linux
- 4 minutes for ppc64le + Linux

Therefore caching the Nim installation reduces the build durations by
about the above amounts when there is no update for the given Nim
`CHANNEL`.

Example cache contents (for Linux on amd64 with Nim 1.0.4):
4.7M  ./bin/nim-stable/.git
5.0M  ./bin/nim-stable/bin/nim
1.1M  ./bin/nim-stable/bin/nimble
39K   ./bin/nim-stable/config
4.3M  ./bin/nim-stable/lib
----
15M   ./bin/nim-stable

This produces a Travis-reported compressed cache size of about 8 MB per
job in the build matrix.

Decisions:
- Show the cache contents, with apparent file sizes.
- Reduce the disk usage of the Nim installation and the `.git` folder to
  reduce the cache size, and do that within the install script, not
  in `before_cache`.
- Don't bother to reduce the cache size further. It would be easy to
  cache only the `.git` directory and the Nim/Nimble binaries (by
  removing the `lib` directory before caching and restoring it each run
  via `git restore` or similar).