Chore(toml): Add `react`, with all tests commented out
This is the only exercise that is in the category of "has canonical
data, but is not up-to-date with the upstream tests". For now, let's add
its `tests.toml` file with all the UUIDs commented out.

Reasons:
- `canonical_data_syncer` will still perceive that all the `react` tests
  are missing, and it reminds us to add them.
- The `git diff` later will be slightly more obvious.

We do not set the tests to `false`, as that implies a decision of "not
implemented" rather than "the existing tests should be updated to comply
with the current canonical data".