CI: Check nim track UUIDs are valid and unique Exercism-wide
Every Exercism track has a `config.json` file that, amongst other
things, specifies the track's exercises. Each exercise must have a UUID.

This commit adds `check_uuids.nim` and a GitHub Actions workflow to run
it during CI. This ensures that every UUID on the Nim track:
- is a valid version 4 UUID in the lowercase and hyphen-separated form
- is unique within the track
- does not exist on another Exercism track
- does not exist in 'exercism/problem-specifications'

The `check_uuids` binary also has a command-line interface that allows:
- specifying the path to the `config.json` file
- disabling the checks that require a network connection

The workflow does not build the binary (or even install Nim) if there
is a cached binary that was built using all of:
- the same source code
- the same version of Ubuntu
- the latest stable version of Nim
- the same versions of GCC and OpenSSL as that installed

In the future, we will probably add most or all of this commit's
uuid-checking functionality to `configlet`. At that time we can remove
the files added by this commit.