exercises(practice): create `instructions.md` from prob-specs descriptions
Practice Exercises in Exercism v3 will be structured differently. One
such change is that there won't be a (generated) README.md file anymore.
This file will be replaced by `.docs/instructions.md` and optionally a
`.docs/introduction.md` file. These files are copies of the information
defined for their corresponding problem-specifications exercise (if
any).

The `.docs/instructions.md` file is a copy of the `instructions.md` file
of the problem-specifications exercise, or if that does not exists, its
`description.md` file.

An upcoming version of `configlet` will add support for doing this
syncing automatically.

See:
- https://github.com/exercism/docs/blob/main/anatomy/tracks/practice-exercises.md
- https://github.com/exercism/configlet