exercises(practice): sync source information from prob-specs
Source information for Practice Exercises will be stored in their
`.meta/config.json` file, in the (optional) `source` and/or `source_url`
properties. This commit adds the source information from the
corresponding problem-specifications exercise to the `.meta/config.json`
file.

An upcoming version of `configlet` will add support for doing this
syncing automatically.

Exercism v3 will automatically include the source information in the
README.md file that is sent via the CLI using the source properties in
the `.meta/config.json` file.

See:
- https://github.com/exercism/docs/blob/main/anatomy/tracks/practice-exercises.md
- https://github.com/exercism/configlet