exercises(practice): move `example.nim` files into `.meta/`
The example implementation file for each Practice Exercise should be
located in the exercise's `.meta` subdirectory. This is consistent with
Concept Exercises, and has an added benefit: we do not need any extra
configuration to ensure that the student does not receive the example
file (as files in the `.meta` subdirectory are never downloaded).

See the spec:
- https://github.com/exercism/docs/blob/main/anatomy/tracks/practice-exercises.md#file-example-implementation