exercises(practice): convert `tests.toml` files to new spec (#288)
Track maintainers found that they wanted to add comments to the
`tests.toml` file to e.g. indicate _why_ a test was not included. But we
were previously using TOML comments to indicate the test name, and
running `configlet sync` would re-generate the entire file without any
manually added comments.

This commit updates our `tests.toml` files to conform to a new
specification that supports documenting such decisions in the file
itself. Each test case is now a separate table with a `description` and
`include` field. A missing `include` field will imply `include = true`.

A track is now free to add additional fields, like a `comment` field,
but also anything they feel might be useful.

Note that the configlet (as of 4.0.0-alpha.10) does not yet support this
new format.

For more background and discussion on the new format, see:
- https://github.com/exercism/configlet/issues/186

Co-authored-by: ee7 <45465154+ee7@users.noreply.github.com>