config.json: add `test_runner.average_run_time` (#301)
This commit adds a new `test_runner.average_run_time` property to the
track `config.json` file.

The purpose of this property is to allow the website to show a progress
bar while the test runner runs. The value should be a float with one
digit after the decimal point, representing the average run time in
seconds.

We have set the value to an average execution time of the test runner
for 5 runs of the `leap` Practice Exercise. Clearly, the execution time
will differ between exercises and between solutions, so this should be
seen as a general indicator.

This is mostly a stop-gap solution for now. We'll revisit this in 6
months time or so.

For more information, see:
- https://github.com/exercism/docs/pull/130