exercises(practice): add authors and contributors (#293)
For each Practice Exercise, we've looked at the commit history of its
files to see which commits touched that Practice Exercise.

We used each commit that we could associate with a Practice Exercise to
determine authorship/contributorship.

To determine authors:
1. Find the Git commit author of the oldest commit linked to that
   Practice Exercise
2. Find the GitHub username for the Git commit author of that commit
3. Add the GitHub username of the Git commit author to the `authors` key
   in the `.meta/config.json` file

To determine contributors:
1. Find the Git commit authors and any co-authors of all but the oldest
   commit linked to that Practice Exercise. If there is only one commit,
   there won't be any contributors.
2. Exclude the Git commit author and any co-authors of the oldest commit
   from the list of Git commit authors (an author is _not_ also a
   contributor)
3. Find the GitHub usernames for the Git commit authors and any
   co-authors of those commits
4. Add the GitHub usernames of the Git commit authors and any co-authors
   to the `contributor` key in the `.meta/config.json` file

We used the GitHub GraphQL API to find the username of a commit author
or any co-authors. In some cases though, a username cannot be found for
a commit (e.g. due to the user account no longer existing), in which
case the commit was skipped.

Handling renames:
There are a small number of Practice Exercises that might have been
renamed at some point. You can ask Git to "follow" a file over its
renames using `git log --follow <file>`, which will also return commits
made before renames.

Unfortunately, Git does not store renames, it just stores the contents
of the renamed files and tries to guess if a file was renamed by looking
at its contents. This _can_ (and will) lead to false positives, where
Git will think a file has been renamed whereas it hasn't.

As we don't want to have incorrect authors/contributors for
exercises, we're ignoring renames. The only exception to this are known
exercise renames:
- `bracket-push` was renamed to `matching-brackets`
- `retree` was renamed to `satellite`
- `resistor-colors` was renamed to `resistor-color-duo`
- `kindergarden-garden` was renamed to `kindergarten-garden`

Exclusions:
There are some commits that we skipped over, and therefore didn't
influence the authors/contributors list:
- Commits authored by `dependabot[bot]`, `dependabot-preview[bot]` or
  `github-actions[bot]`
- Bulk update PRs made by `ErikSchierboom` or `kytrinyx` to update the
  track

ee7 then made some manual tweaks. See #293 for more details.

Co-authored-by: ee7 <45465154+ee7@users.noreply.github.com>