Feature: add generator for exercise tests
This add a generator for exercise test files based on the cannoical data.
This is currently not extensible and there is a lot to be done to improve
the output for objects and the like. It currently just assumes a basic
approach and outputs a tuple for every JSON output of the expected or
input object.

Can be ran with the command:
`nim c -d:slug=<slugname> generator.nim`

The problem specifications repo is assumed to have the same parent
folder as the root of this repo. This can be changed with the
`-d:probSpecExercisesDir=<../problem-specifications/exercises>` switch
when compiling the file.

Future Directions:
1. use the compiler api to not have to recompile for every exercise
2. allow exercise specific configuration for code generation