exercises(protein-translation): update tests and example (#364)
Sync the tests, and make the required change to the example.

The example implementation already had some input validation, even
though it was not required. However, we must modify the example because
the final new test:

    test "Incomplete RNA sequence can translate if valid until a STOP codon":
      check translate("UUCUUCUAAUGGU") == @["Phenylalanine", "Phenylalanine"]

means that `translate` should not immediately raise an exception when
the input length is not divisible by 3.