.github(workflows): move to `iffy/install-nim` 4.4.0
The `alaviss/setup-nim` action cannot [1] install a specific stable
Nim release, only the latest release on a specific branch from the Nim
nightlies.

For robustness, we want to pin the Nim versions that we use in CI. This
commit prepares us for doing so, by switching to the `iffy/install-nim`
action [2] for the non-devel builds.

Therefore, with this commit we:

- use the latest stable release binary, rather than the latest nightly
  on the stable release branch.

- use the latest 1.0 binary (1.0.10), rather than the latest nightly on
  the `version-1-0` branch.

- keep using the latest nightly on the `devel` branch, so that we don't
  have to compile Nim.

[1] https://github.com/alaviss/setup-nim/blob/f81f2a6d1505/setup.sh
[2] https://github.com/iffy/install-nim

Refs: 422
Refs: 427