.github(workflows): put `1.6.8` in matrix
Previously it was unclear that `nim: stable` was not automatically
updated when there was a new Nim release. Furthermore, this commit
improves consistency with the existing version pinning for 1.0:

    nim: '1.0.10'

However, this commit does remove the single source of truth for the
pinned stable Nim version. There are ways to get it back again, but it
muddies the logic. This commit is at least maximally explicit in the
matrix.