.github: add autoresponder for pausing community contributions (#444)
We're going to take a step back and redesign the volunteering model for
Exercism. Please see the forum post [1] for context.

This commit adds an autoresponder that runs when an issue or PR is
opened. If the person opening the issue is not a member of the Exercism
organization, the autoresponder posts a comment and closes the issue. In
the comment, the author is directed to discuss the issue in the forum.

If the discussion in the forum results in the issue/PR being approved, a
maintainer or staff member will reopen it.

[1] https://forum.exercism.org/t/freeing-our-maintainers-exercism-wide-changes-to-track-repositories/1109