exercises(allergies): overhaul to use enum (#434)
This exercise is intended to practice bit manipulation. Make it somewhat
more idiomatic by avoiding the seq of strings. Of course, for Nim code
in the real world, we should use the built-in `set` type.

The overall design of this exercise hadn't changed since the initial
implementation in 7e55686487ae ("Adds allergies exercise", 2018-04-17).

Closes: #404