exercises(gigasecond): use dateTime, not deprecated initDateTime
Before this commit, compiling the tests for `gigasecond` with Nim 1.6.0
or newer would produce 10 deprecation warnings, starting with:

    test_gigasecond.nim(4, 7) template/generic instantiation of `suite` from here
    test_gigasecond.nim(5, 8) template/generic instantiation of `test` from here
    test_gigasecond.nim(6, 20) Warning: use `dateTime`; initDateTime is deprecated [Deprecated]
    [...]

This is because Nim 1.6.0 added a `dateTime` proc that uses the ISO 8601
order, and deprecated the `initDateTime` proc [1][2].

With commit 1ac18dfba497 (".github, docs: bump minimum Nim version from
1.0.0 to 1.6.0", 2023-01-12) the Nim track now officially only supports
Nim 1.6.0 or newer. So let's stop supporting older Nim in this exercise.

[1] https://github.com/nim-lang/Nim/blob/1e52423774e8/changelogs/changelog_1_6_0.md
[2] https://github.com/nim-lang/Nim/commit/4c1202972abd

Closes: #343