exercises(darts): use separate x and y parameters
I think that the procedure in this exercise should take either two
separate parameters, or an object (named e.g. `Coord` or `Point`). Using
a tuple parameter feels like the worst of both worlds.

Let's go with two parameters for now. A common way to solve this
exercise is with `hypot`, which also takes separate parameters.