exercises(darts): use separate parameters, not tuple (#476)
The procedure in this exercise should probably take either two separate
parameters, or an object (named e.g. `Coord` or `Point`). Using a tuple
parameter feels like the worst option.

Let's go with two parameters for now. A common way to solve this
exercise is with `math.hypot`, which also takes separate parameters.