exercises: examples: refactor examples for 40 exercises (#540)
Many example solutions were essentially unchanged since the original
implementation, and some were particularly non-idiomatic. However,
refactoring them was a low priority because they aren't user-facing -
they just needed to pass the tests.

For example, there was still a reference to Nimrod, the old name of the
language. The renaming to Nim happened more than 9 years ago [1][2].

For the example solutions in this repo, there's tension between factors
like:

1. Maintainability: the shortest/simplest code that passes the tests.

2. Being idiomatic: for example, code that a team of fluent Nim
   developers would happily approve in a real-world code review.

3. Being context-aware: a solution that we'd be happy for a student to
   submit.

4. Speed: a hyper-optimized solution that may overfit to the given
   test cases, use SIMD, etc., at the cost of greater complexity.

5. Test coverage: a solution that tries to handle edge cases that aren't
   currently tested.

Make the example solutions more idiomatic, optimizing for performance
where the increase in complexity is low. The larger changes tend to be
for exercises that were added before I began contributing to this repo.

The tests in this repo intentionally support only one `example.nim` per
exercise. If we wanted to test a wider range of solutions, I think it'd
be better to add approaches [3] or articles [4].

With this commit, Nim 2.0.0 still produces various Uninit and ProveInit
warnings, like:

    /foo/largest_series_product.nim(31, 20) Warning: use explicit initialization of 'result' for clarity [Uninit]
    /foo/largest_series_product.nim(5, 3) Warning: Cannot prove that 'result' is initialized. This will become a compile time error in the future. [ProveInit]

where the warnings would be resolved by adding code like:

    result = 0

but let's defer such small changes to a later commit.

Closes: #408

[1] https://github.com/nim-lang/Nim/commit/dbf9117c562601d6b571d964a6cb152e95fe8c4e
[2] https://nim-lang.org/blog/2014/10/19/version-096-released.html
[3] https://exercism.org/docs/building/tracks/approaches
[4] https://exercism.org/docs/building/tracks/articles